Replace float -> decimal to address SQL warnings

+++ sql/food_purchases.sql
@@ -1,7 +1,7 @@
 create table food_purchases(
   `purchase_date` date not null,
   `food_type` varchar(200),
-  `quantity` float(11,2)
+  `quantity` decimal(11,2)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 insert into food_purchases(purchase_date, food_type, quantity) values
